[FEATURE] Stabilizing participation and view of competition type text

+++ Classes/Controller/ParticipationController.php
@@ -1,7 +1,6 @@
 <?php
 namespace Visol\EasyvoteCompetition\Controller;
 
-
 /***************************************************************
  *
  *  Copyright notice
@@ -31,15 +30,7 @@ namespace Visol\EasyvoteCompetition\Controller;
 /**
  * ParticipationController
  */
-class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController {
-
-	/**
-	 * participationRepository
-	 *
-	 * @var \Visol\EasyvoteCompetition\Domain\Repository\ParticipationRepository
-	 * @inject
-	 */
-	protected $participationRepository = NULL;
+class ParticipationController extends AbstractController {
 
 	/**
 	 * action list
@@ -49,6 +40,45 @@ class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCo
 	public function listAction() {
 		$participations = $this->participationRepository->findAll();
 		$this->view->assign('participations', $participations);
+	}
+
+	/**
+	 * @param \Visol\EasyvoteCompetition\Domain\Model\Competition $competition
+	 */
+	public function listByCompetitionAction(\Visol\EasyvoteCompetition\Domain\Model\Competition $competition) {
+		$participations = $this->participationRepository->findByCompetition($competition);
+		$this->view->assign('participations', $participations);
+
+		$communityUser = $this->competitionService->getCurrentCommunityUser();
+		$this->view->assign('communityUser', $communityUser);
+	}
+
+	/**
+	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $participation
+	 * @return bool
+	 */
+	public function addVoteAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
+		$communityUser = $this->competitionService->getCurrentCommunityUser();
+		if ($communityUser instanceof \Visol\Easyvote\Domain\Model\CommunityUser) {
+			if ($this->competitionService->userCanVoteForParticiation($participation->getUid())) {
+				/** @var \Visol\EasyvoteCompetition\Domain\Model\Vote $vote */
+				$vote = $this->objectManager->get('Visol\EasyvoteCompetition\Domain\Model\Vote');
+				$vote->setParticipation($participation);
+				$vote->setCommunityUser($communityUser);
+				$vote->setCompetition($participation->getCompetition());
+				$this->voteRepository->add($vote);
+				$participation->addVote($vote);
+				$this->participationRepository->update($participation);
+				$this->persistenceManager->persistAll();
+				// update cached votes and cached rank
+				$this->participationRepository->updateCachedVotesAndCachedRankForCompetition($participation->getCompetition());
+				return TRUE;
+			} else {
+				// @todo access denied
+				return FALSE;
+			}
+		}
+		return FALSE;
 	}
 
 	/**
@@ -62,37 +92,68 @@ class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCo
 	}
 
 	/**
-	 * action new
-	 *
-	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $newParticipation
-	 * @ignorevalidation $newParticipation
-	 * @return void
-	 */
-	public function newAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $newParticipation = NULL) {
-		$this->view->assign('newParticipation', $newParticipation);
-	}
-
-	/**
-	 * action create
-	 *
-	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $newParticipation
-	 * @return void
-	 */
-	public function createAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $newParticipation) {
-		$this->addFlashMessage('The object was created. Please be aware that this action is publicly accessible unless you implement an access check. See <a href="http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain" target="_blank">Wiki</a>', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
-		$this->participationRepository->add($newParticipation);
-		$this->redirect('list');
-	}
-
-	/**
 	 * action edit
 	 *
+	 * @param \Visol\EasyvoteCompetition\Domain\Model\Competition $competition
 	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $participation
 	 * @ignorevalidation $participation
 	 * @return void
 	 */
-	public function editAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
-		$this->view->assign('participation', $participation);
+	public function editAction(\Visol\EasyvoteCompetition\Domain\Model\Competition $competition = NULL, \Visol\EasyvoteCompetition\Domain\Model\Participation $participation = NULL) {
+		// @TODO check for passed optional participation and if it is allowed to edit it --> possible admin functionality
+		if ($communityUser = $this->getLoggedInUser()) {
+			$mayBeEdited = TRUE;
+			$isNewParticipation = FALSE;
+			$existingParticipation = $this->participationRepository->findOneByCompetitionAndCommunityUser($competition, $communityUser);
+			if (count($existingParticipation)) {
+				// edit the existing participation
+				$participation = $existingParticipation;
+				if (count($participation->getVotes()) > 0) {
+					// participation already has votes --> may not be edited
+					$mayBeEdited = FALSE;
+				}
+				if ($participation->getVotingEnabled() === 2) {
+					// participation was verified
+					$mayBeEdited = FALSE;
+				}
+			} else {
+				// generate an empty new participation that is disabled and save it
+				$isNewParticipation = TRUE;
+				/** @var \Visol\EasyvoteCompetition\Domain\Model\Participation $participation */
+				$participation = $this->objectManager->get('\Visol\EasyvoteCompetition\Domain\Model\Participation');
+				$participation->setCompetition($competition);
+				$participation->setCommunityUser($communityUser);
+				$participation->setDisabled(TRUE);
+				$participation->setVotingEnabled(FALSE);
+				$this->participationRepository->add($participation);
+				$this->persistenceManager->persistAll();
+			}
+			$this->view->assign('isNewParticipation', $isNewParticipation);
+			$this->view->assign('mayBeEdited', $mayBeEdited);
+			$this->view->assign('participation', $participation);
+		} else {
+			// @todo access denied
+		}
+	}
+
+	/**
+	 * action update
+	 *
+	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $participation
+	 * @return void
+	 */
+	public function updatePreviewAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
+		if ($communityUser = $this->getLoggedInUser()) {
+			if ($communityUser === $participation->getCommunityUser()) {
+				$this->participationRepository->update($participation);
+				$this->view->assign('participation', $participation);
+				$this->participationRepository->updateCachedVotesAndCachedRankForCompetition($participation->getCompetition());
+			} else {
+				// @todo access denied
+			}
+		} else {
+			// @todo access denied
+		}
 	}
 
 	/**
@@ -102,9 +163,31 @@ class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCo
 	 * @return void
 	 */
 	public function updateAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
-		$this->addFlashMessage('The object was updated. Please be aware that this action is publicly accessible unless you implement an access check. See <a href="http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain" target="_blank">Wiki</a>', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
 		$this->participationRepository->update($participation);
 		$this->redirect('list');
+	}
+
+	/**
+	 * action publish
+	 *
+	 * @param \Visol\EasyvoteCompetition\Domain\Model\Participation $participation
+	 * @return void
+	 */
+	public function publishAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
+		if ($communityUser = $this->getLoggedInUser()) {
+			if ($communityUser === $participation->getCommunityUser()) {
+				$participation->setVotingEnabled(TRUE);
+				$participation->setDisabled(FALSE);
+				$this->participationRepository->update($participation);
+				$this->persistenceManager->persistAll();
+				$this->participationRepository->updateCachedVotesAndCachedRankForCompetition($participation->getCompetition());
+				$this->redirect('show', 'Competition', $this->request->getControllerExtensionName(), array('competition' => $participation->getCompetition(), 'participation' => $participation, 'openSharer' => 1, 'no_cache' => 1), $this->settings['competitionHomePid']);
+			} else {
+				// @todo access denied
+			}
+		} else {
+			// @todo access denied
+		}
 	}
 
 	/**
@@ -114,9 +197,19 @@ class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCo
 	 * @return void
 	 */
 	public function deleteAction(\Visol\EasyvoteCompetition\Domain\Model\Participation $participation) {
-		$this->addFlashMessage('The object was deleted. Please be aware that this action is publicly accessible unless you implement an access check. See <a href="http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain" target="_blank">Wiki</a>', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
-		$this->participationRepository->remove($participation);
-		$this->redirect('list');
+		if ($communityUser = $this->getLoggedInUser()) {
+			if ($communityUser === $participation->getCommunityUser()) {
+				$competition = $participation->getCompetition();
+				$this->participationRepository->remove($participation);
+				$this->persistenceManager->persistAll();
+				$this->participationRepository->updateCachedVotesAndCachedRankForCompetition($participation->getCompetition());
+				$this->redirect('show', 'Competition', $this->request->getControllerExtensionName(), array('competition' => $competition), $this->settings['competitionHomePid']);
+			} else {
+				// @todo access denied
+			}
+		} else {
+			// @todo access denied
+		}
 	}
 
 	/**
@@ -125,7 +218,6 @@ class ParticipationController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionCo
 	 * @return void
 	 */
 	public function listForCommunityUserAction() {
-		
 	}
 
 }
